Add files via upload
The Java program represents a simple petrol pump management system. It begins by displaying a welcome message and then presents a menu of options for the user to select from, including filling petrol or diesel, direct filling options, and accessing settings. The user is prompted to enter their choice, and based on their selection, the program executes the corresponding functionality such as filling petrol or diesel based on the entered amount in rupees, displaying the cost in terms of liters and price per liter. Additionally, there's a setting option available to the user, which displays different settings options. The program utilizes Java's Scanner class for user input and Thread.sleep() for adding delays, and it clears the screen for a better user interface. Overall, it provides a basic interface for managing petrol pump operations, including fuel filling and settings access.